fixed SyncClient
