+++ src/inmanta/util.py
@@ -733,3 +733,18 @@ async def join_threadpools(threadpools: List[ThreadPoolExecutor]) -> None:
         await future
     finally:
         thread.join()
+
+
+def ensure_event_loop() -> asyncio.AbstractEventLoop:
+    """
+    Returns the event loop for this thread. Creates a new one if none exists yet and registers it with asyncio's active event
+    loop policy.
+    """
+    try:
+        # nothing needs to be done if this thread already has an event loop
+        return asyncio.get_event_loop()
+    except RuntimeError:
+        # asyncio.set_event_loop sets the event loop for this thread only
+        new_loop = asyncio.new_event_loop()
+        asyncio.set_event_loop(new_loop)
+        return new_loop
